+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.22.0] - 2024-11-26
+
+### Changed
+
+- Single marker lazy loading for very large ROIs (both large dimensions + many markers): includes
+spatial datasets and extended to mcd & tiff
+
 ## [0.21.0] - 2024-11-21
 
 ### Added